feat: move cone scan pipeline to separate file
