+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/autonomous/ConeParking.kt
@@ -16,9 +16,8 @@ import org.openftc.easyopencv.OpenCvInternalCamera
 import org.openftc.easyopencv.OpenCvPipeline
 import kotlin.math.PI
 
-abstract class ConeParking : LinearOpMode() {
-    abstract val direction: Double
-
+@Autonomous(name = "Cone Parking")
+class ConeParking : LinearOpMode() {
     private val config = Config(runMode = RunMode.Autonomous)
 
     private var ctx: Context? = null
@@ -52,10 +51,14 @@ abstract class ConeParking : LinearOpMode() {
         this.distance_sensor.initPlugin(this.ctx!!)
         this.distance_sensor.init()
 
-        val cameraMonitorViewId = hardwareMap.appContext.resources.getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.packageName)
+        val cameraMonitorViewId = hardwareMap.appContext.resources.getIdentifier(
+            "cameraMonitorViewId",
+            "id",
+            hardwareMap.appContext.packageName
+        )
         val camera = OpenCvCameraFactory.getInstance().createInternalCamera(
-                OpenCvInternalCamera.CameraDirection.BACK,
-                cameraMonitorViewId
+            OpenCvInternalCamera.CameraDirection.BACK,
+            cameraMonitorViewId
         )
 
         val pipeline = ConeScanPipeline()
@@ -77,14 +80,13 @@ abstract class ConeParking : LinearOpMode() {
 
         val runtime = ElapsedTime()
 
-        // TODO: SCAN COLOR AND STORE IT IN A VARIABLE FOR LATER
-
         val coneColor = pipeline.output
 
         camera.stopStreaming()
         camera.closeCameraDevice()
 
         runtime.reset()
+
         if (coneColor == ConeScanPipeline.Color.Green) {
             while (runtime.seconds() < 2.0 && opModeIsActive()) {
                 wheels.powerDirection(2 * PI / 3, 0.5)
@@ -100,48 +102,12 @@ abstract class ConeParking : LinearOpMode() {
             while (runtime.seconds() < 1.25 && opModeIsActive()) {
                 wheels.powerDirection(PI / 6, 0.5)
             }; wheels.stop()
-            while (runtime.seconds()< 2.55 && opModeIsActive()) {
+            while (runtime.seconds() < 2.55 && opModeIsActive()) {
                 wheels.stop()
             }
             while (runtime.seconds() < 4.0 && opModeIsActive()) {
                 wheels.powerDirection((2 * PI / 3) - 0.1, 0.5)
             }
-        }
-    }
-
-
-    @Autonomous(name = "ConeParking")
-    class ConePaking : ConeParking() {
-        override val direction = 0.0
-
-    }
-
-    class ConeScanPipeline : OpenCvPipeline() {
-        enum class Color {
-            Red,
-            Green,
-            Blue,
-        }
-
-        var output = Color.Blue
-            private set
-
-        private var contrasted = Mat()
-
-        override fun processFrame(input: Mat?): Mat {
-            input!!.convertTo(contrasted, -1, 0.6)
-
-            val p = contrasted.get(contrasted.rows() / 2, contrasted.cols() / 2)
-
-            output = if (p[0] >= p[1] && p[0] >= p[2]) {
-                Color.Red
-            } else if (p[1] >= p[0] && p[1] >= p[2]) {
-                Color.Green
-            } else {
-                Color.Blue
-            }
-
-            return contrasted
         }
     }
 }
